+++ R/dplyr_compatibility.R
@@ -23,3 +23,16 @@ setMethod(
     dplyr::filter(object, ...)
   }
 )
+
+# rename-data.frame ----
+
+#' @rdname rename
+setMethod(
+  f = "rename",
+  signature = signature(x = "data.frame"),
+  definition = function(x, ...) {
+    # We just forward to the `dplyr` generic function, so that the user
+    # can still rename data frames.
+    dplyr::rename(x, ...)
+  }
+)