+++ R/draw_boxplot.R
@@ -2,19 +2,20 @@
 #'
 #' @description `r lifecycle::badge("experimental")`
 #'
-#' This produces a boxplot of the gene expression values of a gene for a
-#' sample variable.
+#' This produces boxplots of the gene expression values of a single gene, multiple genes or
+#' a gene signature.
 #'
 #' @param object (`AnyHermesData`)\cr input.
 #' @param assay_name (`string`)\cr selects assay from input for the y-axis.
-#' @param genes (`character`)\cr gene ID(s) for which to produce boxplots.
+#' @param genes (`GeneSpec`)\cr for which genes or which gene signature to produce boxplots.
 #' @param x_var (`string` or `NULL`)\cr optional stratifying variable for the x-axis,
 #'   taken from input sample variables.
 #' @param color_var (`string` or `NULL`)\cr optional color variable, taken from
 #'   input sample variables.
 #' @param facet_var (`string` or `NULL`)\cr optional faceting variable, taken
 #'   from input sample variables.
-#' @param jitter (`flag`)\cr whether to add jittered original data points or not.
+#' @param violin (`flag`)\cr whether to draw a violin plot instead of a boxplot.
+#' @param jitter (`flag`)\cr whether to add jittered original data points.
 #'
 #' @return The `ggplot` boxplot.
 #'
@@ -25,7 +26,8 @@
 #' draw_boxplot(
 #'   object,
 #'   assay_name = "counts",
-#'   genes = genes(object)[2]
+#'   genes = gene_spec(genes(object)[1]),
+#'   violin = TRUE
 #' )
 #'
 #' object2 <- object %>%
@@ -36,7 +38,7 @@
 #'   object2,
 #'   assay_name = "tpm",
 #'   x_var = "SEX",
-#'   genes = genes(object2)[20],
+#'   genes = gene_spec(genes(object2)[1:10], fun = colMeans),
 #'   facet_var = "RACE",
 #'   color_var = "AGE18",
 #'   jitter = TRUE
@@ -46,16 +48,9 @@
 #'   object,
 #'   assay_name = "counts",
 #'   x_var = "SEX",
-#'   genes = genes(object)[2],
-#'   jitter = TRUE
-#' )
-#'
-#' draw_boxplot(
-#'   object,
-#'   assay_name = "counts",
-#'   x_var = "SEX",
-#'   genes = genes(object)[1:2],
-#'   facet_var = "RACE"
+#'   genes = gene_spec(genes(object)[1:3]),
+#'   jitter = TRUE,
+#'   facet_var = "AGE18"
 #' )
 draw_boxplot <- function(object,
                          assay_name,
@@ -63,18 +58,20 @@ draw_boxplot <- function(object,
                          x_var = NULL,
                          color_var = NULL,
                          facet_var = NULL,
+                         violin = FALSE,
                          jitter = FALSE) {
   assert_class(object, "AnyHermesData")
   assert_string(assay_name)
-  assert_character(genes, any.missing = FALSE, unique = TRUE)
+  assert_class(genes, "GeneSpec")
   assert_string(x_var, null.ok = TRUE)
   assert_string(color_var, null.ok = TRUE)
   assert_string(facet_var, null.ok = TRUE)
+  assert_flag(violin)
   assert_flag(jitter)
 
   assay_matrix <- assay(object, assay_name)
   col_data <- colData(object)
-  assert_names(rownames(assay_matrix), must.include = genes)
+  assert_names(rownames(assay_matrix), must.include = genes$get_genes())
 
   x <- if (!is.null(x_var)) {
     assert_names(names(col_data), must.include = x_var)
@@ -83,10 +80,15 @@ draw_boxplot <- function(object,
     factor(0)
   }
 
+  fill_vals <- if (genes$returns_vector()) {
+    genes$get_label()
+  } else {
+    genes$get_gene_labels()
+  }
   df <- data.frame(
     x = x,
-    y = as.numeric(t(assay_matrix[genes, , drop = FALSE])),
-    fill = factor(rep(genes, each = ncol(assay_matrix)))
+    y = as.numeric(t(genes$extract(assay_matrix))),
+    fill = factor(rep(fill_vals, each = ncol(assay_matrix)))
   )
 
   if (!is.null(facet_var)) {
@@ -104,11 +106,19 @@ draw_boxplot <- function(object,
     aes(group = .data$fill)
   }
   p <- ggplot(df, aes(x = .data$x, y = .data$y, fill = .data$fill)) +
-    geom_boxplot(outlier.shape = ifelse(jitter, NA, 19)) +
-    stat_boxplot(geom = "errorbar") +
+    labs(x = genes$get_label())
+  p <- if (!violin) {
+    p +
+      geom_boxplot(outlier.shape = ifelse(jitter, NA, 19)) +
+      stat_boxplot(geom = "errorbar")
+  } else {
+    p +
+      geom_violin(draw_quantiles = c(0.75, 0.5, 0.25))
+  }
+  p <- p +
     geom_point(
       mapping = point_aes,
-      position = position_jitterdodge(jitter.width = jitter_width),
+      position = position_jitterdodge(jitter.width = jitter_width)
     ) +
     labs(x = x_var, y = assay_name, fill = "Gene")
   if (is.null(x_var)) {